Add skype user to irc group
When running Skype container from Fedora, the gid for "video" group is
not the same as in Debian; in fact, it correspond to the gid for "irc"
group in Debian.

So in the container, /dev/video0 belongs to "irc" group, and thus skype
can't use it.

So a quick fix is to add skype to irc group, so it can use it without
any problem.